box: allow printable and only printable characters in identifiers
Before this patch, we would only allow alphabetical characters
plus underscore in identifier names. And we did not treat
all identifiers the same way: column names were not checked
at all.

SQL ANSI ISO allow delimited identifiers cantain any character
from source language character set.

After this patch, checks for allowed characters in identifier
names follow the same ruls for all identifiers: column names,
function names, user names, space names, index names.

In other words, this patch makes tarantool itentifier rules closer
to ANSI ones.

Closes #2914

+++ src/trivia/util.h
@@ -499,7 +499,7 @@ tt_static_buf(void)
  * Return a null-terminated string for \a str of length \a len
  */
 static inline const char *
-tt_cstr(const char *str, uint32_t len)
+tt_cstr(const char *str, size_t len)
 {
 	char *buf = tt_static_buf();
 	len = MIN(len, TT_STATIC_BUF_LEN - 1);
